Fix. Avoided a few precision warnings.

+++ trunk/nebula2/code/nebula2/src/gfx2/nd3d9server_render.cc
@@ -35,10 +35,10 @@ nD3D9Server::SetViewport(nViewport& vp)
     if (this->d3d9Device)
     {
         static D3DVIEWPORT9 dvp;
-        dvp.X = vp.x;
-        dvp.Y = vp.y;
-        dvp.Width = vp.width;
-        dvp.Height = vp.height;
+        dvp.X = (DWORD)vp.x;
+        dvp.Y = (DWORD)vp.y;
+        dvp.Width = (DWORD)vp.width;
+        dvp.Height = (DWORD)vp.height;
         dvp.MinZ = vp.nearz;
         dvp.MaxZ = vp.farz;
         this->d3d9Device->SetViewport(&dvp);
